Add files via upload
Using this app, one can create new tasks, edit it, mark it as whether it is completed or incomplete. Once you complete the task, you can either mark it as complete or delete it. You can create multiple sets of tasks which are locally stored and can be edited anytime.